Extract `sell_complete_set_tests`
